logging

+++ nestjs-task-management/src/tasks/tasks.service.ts
@@ -7,6 +7,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -16,13 +17,19 @@ export class TasksService {
     ) {}
 
     // get all task
-    async getTasks(filterDto: getTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto)
+    async getTasks(
+        filterDto: getTasksFilterDto,
+        user: User
+        ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user)
     }
 
     // get task by ID
-    async getTaskById(id: number) : Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User
+        ) : Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
         if(!found){
             throw new NotFoundException(`Task with "${id}" not found`);
@@ -33,20 +40,34 @@ export class TasksService {
 
 
     // eksekusi ke database dilakuakn oleh repository
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> { 
-        console.log(createTaskDto);
-        return this.taskRepository.createTask(createTaskDto)
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User
+        ): Promise<Task> { 
+        return this.taskRepository.createTask(createTaskDto, user)
     }
 
     // delete task by ID
-    async deleteTask(id: number) : Promise<void> {
-       const result = await this.taskRepository.delete(id)
+    async deleteTask(
+        id: number,
+        user: User
+        ) : Promise<void> {
+       const result = await this.taskRepository.delete({id, userId: user.id})
+
+       if(result.affected === 0){
+        throw new NotFoundException(`Task with "${id}" not found`);
+    }
+
        console.log(result)
     }
 
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {  
-        const task = await this.getTaskById(id)
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus,
+        user: User
+        ): Promise<Task> {  
+        const task = await this.getTaskById(id, user)
         task.status = status;
         await task.save();
         return task;
